Refactor all code that interfaces with FFXIV memory
