+++ OverlayPlugin.Core/MemoryProcessors/Aggro/AggroMemory61.cs
@@ -1,0 +1,9 @@
+﻿namespace RainbowMage.OverlayPlugin.MemoryProcessors.Aggro
+{
+    class AggroMemory61 : AggroMemory60
+    {
+        public AggroMemory61(TinyIoCContainer container)
+            : base(container)
+        { }
+    }
+}
