+++ OverlayPlugin.Core/MemoryProcessors/Aggro/AggroMemory62.cs
@@ -1,0 +1,9 @@
+﻿namespace RainbowMage.OverlayPlugin.MemoryProcessors.Aggro
+{
+    class AggroMemory62 : AggroMemory61
+    {
+        public AggroMemory62(TinyIoCContainer container)
+            : base(container)
+        { }
+    }
+}
