+++ OverlayPlugin.Core/MemoryProcessors/Combatant/CombatantMemoryManager.cs
@@ -1,0 +1,93 @@
+﻿using System;
+using System.Collections.Generic;
+
+namespace RainbowMage.OverlayPlugin.MemoryProcessors.Combatant
+{
+    public interface ICombatantMemory
+    {
+        bool IsValid();
+        Combatant GetSelfCombatant();
+        Combatant GetCombatantFromAddress(IntPtr address, uint selfCharID);
+        List<Combatant> GetCombatantList();
+    }
+
+    public class CombatantMemoryManager : ICombatantMemory
+    {
+        private readonly TinyIoCContainer container;
+        private readonly FFXIVRepository repository;
+        private CombatantMemory memory = null;
+
+        public CombatantMemoryManager(TinyIoCContainer container)
+        {
+            this.container = container;
+            container.Register(new CombatantMemory60(container));
+            container.Register(new CombatantMemory61(container));
+            container.Register(new CombatantMemory62(container));
+            repository = container.Resolve<FFXIVRepository>();
+        }
+
+        private void FindMemory()
+        {
+            List<CombatantMemory> candidates = new List<CombatantMemory>();
+            // For CN/KR, try the lang-specific candidate first, then fall back to intl
+            if (repository.GetLanguage() == FFXIV_ACT_Plugin.Common.Language.Chinese)
+            {
+                candidates.Add(container.Resolve<CombatantMemory61>());
+            }
+            else if (repository.GetLanguage() == FFXIV_ACT_Plugin.Common.Language.Korean)
+            {
+                candidates.Add(container.Resolve<CombatantMemory60>());
+            }
+            candidates.Add(container.Resolve<CombatantMemory62>());
+
+            foreach (var c in candidates)
+            {
+                if (c.IsValid())
+                {
+                    memory = c;
+                    break;
+                }
+            }
+        }
+
+        public bool IsValid()
+        {
+            if (memory == null)
+            {
+                FindMemory();
+            }
+            if (memory == null || !memory.IsValid())
+            {
+                return false;
+            }
+            return true;
+        }
+
+        public Combatant GetCombatantFromAddress(IntPtr address, uint selfCharID)
+        {
+            if (!IsValid())
+            {
+                return null;
+            }
+            return memory.GetCombatantFromAddress(address, selfCharID);
+        }
+
+        public List<Combatant> GetCombatantList()
+        {
+            if (!IsValid())
+            {
+                return new List<Combatant>();
+            }
+            return memory.GetCombatantList();
+        }
+
+        public Combatant GetSelfCombatant()
+        {
+            if (!IsValid())
+            {
+                return null;
+            }
+            return memory.GetSelfCombatant();
+        }
+    }
+}
