+++ OverlayPlugin.Core/MemoryProcessors/Enmity/EnmityMemory61.cs
@@ -1,0 +1,9 @@
+﻿namespace RainbowMage.OverlayPlugin.MemoryProcessors.Enmity
+{
+    class EnmityMemory61 : EnmityMemory60
+    {
+        public EnmityMemory61(TinyIoCContainer container)
+            : base(container)
+        { }
+    }
+}
