+++ OverlayPlugin.Core/MemoryProcessors/Enmity/EnmityMemory62.cs
@@ -1,0 +1,9 @@
+﻿namespace RainbowMage.OverlayPlugin.MemoryProcessors.Enmity
+{
+    class EnmityMemory62 : EnmityMemory61
+    {
+        public EnmityMemory62(TinyIoCContainer container)
+            : base(container)
+        { }
+    }
+}
