+++ OverlayPlugin.Core/MemoryProcessors/EnmityHud/EnmityHudMemory61.cs
@@ -1,0 +1,9 @@
+﻿namespace RainbowMage.OverlayPlugin.MemoryProcessors.EnmityHud
+{
+    class EnmityHudMemory61 : EnmityHudMemory60
+    {
+        public EnmityHudMemory61(TinyIoCContainer container)
+            : base(container)
+        { }
+    }
+}
