+++ OverlayPlugin.Core/MemoryProcessors/InCombat/InCombatMemory62.cs
@@ -1,0 +1,8 @@
+﻿namespace RainbowMage.OverlayPlugin.MemoryProcessors.InCombat
+{
+    // No changes in 6.2
+    class InCombatMemory62 : InCombatMemory61
+    {
+        public InCombatMemory62(TinyIoCContainer container) : base(container) { }
+    }
+}
