+++ OverlayPlugin.Core/OverlayPlugin.Core.csproj
@@ -91,15 +91,40 @@
     <Compile Include="EventSources\CactbotJSEvents.cs" />
     <Compile Include="EventSources\EnmityEventSource.cs" />
     <Compile Include="Integration\UnstableNewLogLines.cs" />
-    <Compile Include="MemoryProcessors\EnmityMemory50.cs" />
-    <Compile Include="MemoryProcessors\EnmityMemory52.cs" />
-    <Compile Include="MemoryProcessors\EnmityMemory53.cs" />
-    <Compile Include="MemoryProcessors\EnmityMemory62.cs" />
-    <Compile Include="MemoryProcessors\EnmityMemory61.cs" />
-    <Compile Include="MemoryProcessors\EnmityMemory60.cs" />
-    <Compile Include="MemoryProcessors\EnmityMemory55.cs" />
-    <Compile Include="MemoryProcessors\EnmityMemory54.cs" />
-    <Compile Include="MemoryProcessors\EnmityMemoryCommon.cs" />
+    <Compile Include="MemoryProcessors\Aggro\AggroMemory.cs" />
+    <Compile Include="MemoryProcessors\Aggro\AggroMemory60.cs" />
+    <Compile Include="MemoryProcessors\Aggro\AggroMemory61.cs" />
+    <Compile Include="MemoryProcessors\Aggro\AggroMemory62.cs" />
+    <Compile Include="MemoryProcessors\Aggro\AggroMemoryManager.cs" />
+    <Compile Include="MemoryProcessors\Aggro\Common.cs" />
+    <Compile Include="MemoryProcessors\Combatant\Common.cs" />
+    <Compile Include="MemoryProcessors\EnmityHud\Common.cs" />
+    <Compile Include="MemoryProcessors\EnmityHud\EnmityHudMemory.cs" />
+    <Compile Include="MemoryProcessors\EnmityHud\EnmityHudMemory60.cs" />
+    <Compile Include="MemoryProcessors\EnmityHud\EnmityHudMemory61.cs" />
+    <Compile Include="MemoryProcessors\EnmityHud\EnmityHudMemory62.cs" />
+    <Compile Include="MemoryProcessors\EnmityHud\EnmityHudMemoryManager.cs" />
+    <Compile Include="MemoryProcessors\Enmity\Common.cs" />
+    <Compile Include="MemoryProcessors\Enmity\EnmityMemory.cs" />
+    <Compile Include="MemoryProcessors\Enmity\EnmityMemory60.cs" />
+    <Compile Include="MemoryProcessors\Enmity\EnmityMemory61.cs" />
+    <Compile Include="MemoryProcessors\Enmity\EnmityMemory62.cs" />
+    <Compile Include="MemoryProcessors\Enmity\EnmityMemoryManager.cs" />
+    <Compile Include="MemoryProcessors\Combatant\CombatantMemoryManager.cs" />
+    <Compile Include="MemoryProcessors\InCombat\InCombatMemory.cs" />
+    <Compile Include="MemoryProcessors\InCombat\InCombatMemory60.cs" />
+    <Compile Include="MemoryProcessors\InCombat\InCombatMemory61.cs" />
+    <Compile Include="MemoryProcessors\InCombat\InCombatMemory62.cs" />
+    <Compile Include="MemoryProcessors\InCombat\TargetMemoryManager.cs" />
+    <Compile Include="MemoryProcessors\Target\TargetMemoryManager.cs" />
+    <Compile Include="MemoryProcessors\Target\TargetMemory.cs" />
+    <Compile Include="MemoryProcessors\Target\TargetMemory60.cs" />
+    <Compile Include="MemoryProcessors\Target\TargetMemory61.cs" />
+    <Compile Include="MemoryProcessors\Target\TargetMemory62.cs" />
+    <Compile Include="MemoryProcessors\Combatant\CombatantMemory60.cs" />
+    <Compile Include="MemoryProcessors\Combatant\CombatantMemory61.cs" />
+    <Compile Include="MemoryProcessors\Combatant\CombatantMemory62.cs" />
+    <Compile Include="MemoryProcessors\Combatant\CombatantMemory.cs" />
     <Compile Include="NetworkProcessors\FateWatcher.cs" />
     <Compile Include="NetworkProcessors\LineFateControl.cs" />
     <Compile Include="NetworkProcessors\LineCEDirector.cs" />
